requirements.txt
pytest~=7.2.0
selene>=2.0.0b8
allure-pytest==2.12.0
selenium==4.6.1